requirements.txt
google-generativeai
Flask
flask-cors
gunicorn